+++ src/main/java/com/telepathicgrunt/the_bumblezone/blocks/IncenseCandleBase.java
@@ -1,6 +1,9 @@
 package com.telepathicgrunt.the_bumblezone.blocks;
 
 import com.telepathicgrunt.the_bumblezone.blocks.blockentities.IncenseCandleBlockEntity;
+import com.telepathicgrunt.the_bumblezone.entities.mobs.BeeQueenEntity;
+import com.telepathicgrunt.the_bumblezone.items.dispenserbehavior.DispenserItemSetup;
+import com.telepathicgrunt.the_bumblezone.modcompat.ModChecker;
 import com.telepathicgrunt.the_bumblezone.modinit.BzBlockEntities;
 import com.telepathicgrunt.the_bumblezone.modinit.BzBlocks;
 import com.telepathicgrunt.the_bumblezone.modinit.BzCriterias;
@@ -10,6 +13,7 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.nbt.CompoundTag;
+import net.minecraft.network.syncher.EntityDataSerializers;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.tags.FluidTags;
 import net.minecraft.util.RandomSource;
@@ -21,6 +25,7 @@ import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.projectile.Projectile;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
 import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.level.BlockAndTintGetter;
 import net.minecraft.world.level.BlockGetter;
@@ -51,6 +56,8 @@ import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.Vec3;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.VoxelShape;
+import net.minecraftforge.event.entity.player.PlayerEvent;
+import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 
 
 public class IncenseCandleBase extends BaseEntityBlock implements SimpleWaterloggedBlock, SuperCandle {
@@ -291,11 +298,8 @@ public class IncenseCandleBase extends BaseEntityBlock implements SimpleWaterlog
                     int green = Math.max((currentColor >> 8 & 255), 10);
                     int blue = Math.max((currentColor & 255), 5);
                     currentColor = (Math.min(red + 60, 255) << 16) + (Math.min(green + 30, 255) << 8) + Math.min(blue + 25, 255);
-                    return currentColor;
                 }
-                else {
-                    return currentColor;
-                }
+                return currentColor;
             }
         }
         return tintIndex;
@@ -340,5 +344,21 @@ public class IncenseCandleBase extends BaseEntityBlock implements SimpleWaterlog
             return BlockPathTypes.DAMAGE_FIRE;
         }
         return null;
+    }
+
+    public static void multiPotionCandleCrafted(PlayerEvent.ItemCraftedEvent event) {
+        if (event.getEntity() instanceof ServerPlayer serverPlayer && event.getCrafting().is(BzItems.INCENSE_CANDLE.get())) {
+            int containerSize = event.getInventory().getContainerSize();
+            int potionsUsed = 0;
+            for (int i = 0; i < containerSize; i++) {
+                if (event.getInventory().getItem(i).is(Items.POTION)) {
+                    potionsUsed++;
+                }
+            }
+
+            if (potionsUsed >= 2) {
+                BzCriterias.CRAFT_MULTI_POTION_INCENSE_CANDLE_TRIGGER.trigger(serverPlayer);
+            }
+        }
     }
 }