+++ src/main/java/com/telepathicgrunt/the_bumblezone/entities/EntityTeleportationHookup.java
@@ -52,11 +52,11 @@ public class EntityTeleportationHookup {
         //Makes it so player does not get killed for falling into the void
         if (livingEntity.level.dimension().location().equals(Bumblezone.MOD_DIMENSION_ID)) {
             if (livingEntity.getY() < -2) {
-                if(livingEntity instanceof ServerPlayer && livingEntity.fallDistance > 100 && livingEntity.getDeltaMovement().y() < -1) {
-                    BzCriterias.TELEPORT_OUT_OF_BUMBLEZONE_FALL_TRIGGER.trigger((ServerPlayer) livingEntity);
-                }
-
                 if(BzDimensionConfigs.enableExitTeleportation.get()) {
+                    if(livingEntity instanceof ServerPlayer) {
+                        BzCriterias.TELEPORT_OUT_OF_BUMBLEZONE_TRIGGER.trigger((ServerPlayer) livingEntity);
+                    }
+
                     if (livingEntity.getY() < -4) {
                         livingEntity.moveTo(livingEntity.getX(), -4, livingEntity.getZ());
                         livingEntity.absMoveTo(livingEntity.getX(), -4, livingEntity.getZ());
@@ -80,6 +80,10 @@ public class EntityTeleportationHookup {
             }
             else if (livingEntity.getY() > 255) {
                 if(BzDimensionConfigs.enableExitTeleportation.get()) {
+                    if(livingEntity instanceof ServerPlayer) {
+                        BzCriterias.TELEPORT_OUT_OF_BUMBLEZONE_TRIGGER.trigger((ServerPlayer) livingEntity);
+                    }
+
                     if (livingEntity.getY() > 257) {
                         livingEntity.moveTo(livingEntity.getX(), 257, livingEntity.getZ());
                         livingEntity.absMoveTo(livingEntity.getX(), 257, livingEntity.getZ());