+++ src/main/java/com/telepathicgrunt/the_bumblezone/world/dimension/BzSkyProperty.java
@@ -24,11 +24,7 @@ public class BzSkyProperty extends DimensionSpecialEffects {
     public static float REDDISH_FOG_TINT = 0;
 
     /**
-     * Returns fog color
-     * <p>
-     * What I done is made it be based on the day/night cycle so the fog will darken at night but brighten during day.
-     * calculateVanillaSkyPositioning returns a value which is between 0 and 1 for day/night and fogChangeSpeed is the range
-     * that the fog color will cycle between.
+     * Returns fog color based on if player has wrath effect or not
      */
     public Vec3 getFogColor() {
         float colorFactor = 1;
@@ -48,8 +44,8 @@ public class BzSkyProperty extends DimensionSpecialEffects {
             REDDISH_FOG_TINT -= 0.00001f;
         }
 
-        return new Vec3((int)(Math.min(Math.min(0.56f * colorFactor, 0.65f + REDDISH_FOG_TINT)*255, 255)),
-                        ((int)(Math.min(Math.max(Math.min(0.34f * colorFactor, 0.87f) - REDDISH_FOG_TINT * 0.6f, 0)*255, 255))),
+        return new Vec3((int)(Math.min(Math.min(0.54f * colorFactor, 0.65f + REDDISH_FOG_TINT)*255, 255)),
+                        ((int)(Math.min(Math.max(Math.min(0.3f * colorFactor, 0.87f) - REDDISH_FOG_TINT * 0.6f, 0)*255, 255))),
                         ((int)(Math.min(Math.max(Math.min((0.001f * colorFactor) * (colorFactor * colorFactor), 0.9f) - REDDISH_FOG_TINT * 1.9f, 0)*255, 255))));
     }
 }